+++ src/lips.js
@@ -415,6 +415,10 @@
                 }
             }
         });
+        if (!tokens.filter(t => t.match(/^[()]$/)).length && stack.length) {
+            result = result.concat(stack);
+            stack = [];
+        }
         if (stack.length) {
             dump(result);
             throw new Error('Unbalanced parenthesis 2');
@@ -694,14 +698,13 @@
         [['(', 'begin'], 1],
         [['(', 'begin', sexp], 1, notParen],
         [['(', /^let\*?$/, '(', glob, ')'], 1],
-        [['(', /^let\*?$/, new Pattern(['(', glob, ')'], '+')], 1, notParen],
-        [['(', /^let\*?$/, '(', ['(', glob, ')']], 2, notParen],
+        [['(', /^let\*?$/, '(', new Pattern(['(', glob, ')'], '+')], 2, notParen],
         [['(', 'if', /[^()]/], 1],
         [['(', 'if', ['(', glob, ')']], 1],
         [['(', 'if', ['(', glob, ')'], ['(', glob, ')']], 1],
         [['(', glob, ')'], 1],
         [['(', /^define/, '(', glob, ')'], 1],
-        [['(', /^define/, ['(', glob, ')'], sexp], 1],
+        [['(', /^define/, ['(', glob, ')'], sexp], 1, notParen],
         [['(', 'lambda', '(', glob, ')'], 1],
         [['(', 'lambda', ['(', glob, ')'], sexp], 1, notParen]
     ];
@@ -1371,8 +1374,9 @@
     // :: function bind fn with context but it also move all props
     // :: mostly used for Object function
     // ----------------------------------------------------------------------
+    var exludedNames = ['name', 'length', 'caller', 'callee', 'arguments', 'prototype'];
     function filterFnNames(name) {
-        return !['name', 'length', 'caller', 'callee', 'arguments'].includes(name);
+        return !exludedNames.includes(name);
     }
     // ----------------------------------------------------------------------
     function bindWithProps(fn, context) {
@@ -1481,6 +1485,26 @@
                     return unpromise(set(value), loop);
                 }
             })();
+        });
+    }
+    function pararel(name, fn) {
+        return new Macro(name, function(code, { dynamic_scope, error } = {}) {
+            var env = this;
+            if (dynamic_scope === true) {
+                dynamic_scope = this;
+            }
+            var node = code;
+            var results = [];
+            while (node instanceof Pair && !isEmptyList(node)) {
+                results.push(evaluate(node.car, { env, dynamic_scope, error }));
+                node = node.cdr;
+            }
+            var havePromises = results.filter(isPromise).length;
+            if (havePromises) {
+                return Promise.all(results).then(fn.bind(this));
+            } else {
+                return fn.call(this, results);
+            }
         });
     }
     // ----------------------------------------------------------------------
@@ -1690,6 +1714,10 @@
         return LNumber(value);
     };
     // ----------------------------------------------------------------------
+    LNumber.prototype.isFloat = function() {
+        return !!(LNumber.isFloat(this.value) || this.float);
+    };
+    // ----------------------------------------------------------------------
     LNumber.prototype.op = function(op, n) {
         var ops = {
             '*': function(a, b) {
@@ -1723,9 +1751,8 @@
                 return a << b;
             }
         };
-        if (LNumber.isFloat(n) || (n instanceof LNumber &&
-                                   (LNumber.isFloat(n.value) || n.float)) ||
-            (LNumber.isFloat(this.value) || this.float)) {
+        if (LNumber.isFloat(n) || (n instanceof LNumber && n.isFloat()) ||
+            this.isFloat()) {
             var value = n instanceof LNumber ? n.valueOf() : n;
             return LNumber(ops[op](this.valueOf(), value), true);
         }
@@ -2120,6 +2147,7 @@
             if (key instanceof Pair && !(list instanceof Pair)) {
                 throw new Error('First argument to assoc ned to be a key');
             }
+            typecheck('assoc', list, 'pair');
             var node = list;
             while (true) {
                 if (!(node instanceof Pair) || this.get('empty?')(node)) {
@@ -2240,6 +2268,14 @@
              previous values/names when next are evaluated. You can only get them
              from body of let expression.`),
         // ------------------------------------------------------------------
+        'begin*': doc(pararel('begin*', function(values) {
+            return values.pop();
+        }), `(begin* . expr)
+
+             This macro is parallel version of begin. It evaluate each expression and
+             if it's a promise it will evaluate it in parallel and return value
+             of last expression.`),
+        // ------------------------------------------------------------------
         'begin': doc(new Macro('begin', function(code, { dynamic_scope, error }) {
             var arr = this.get('list->array')(code);
             if (dynamic_scope) {
@@ -2281,7 +2317,7 @@
         timer: doc(new Macro('timer', function(code, { dynamic_scope, error } = {}) {
             typecheck('timer', code.car, 'number');
             var env = this;
-            if (dynamic_scope) {
+            if (dynamic_scope === true) {
                 dynamic_scope = this;
             }
             return new Promise((resolve) => {
@@ -2362,6 +2398,7 @@
             Function return current environement.`),
         // ------------------------------------------------------------------
         'eval': doc(function(code, env) {
+            typecheck('eval', code, ['symbol', 'pair', 'array']);
             env = env || this;
             if (code instanceof Symbol) {
                 return env.get(code);
@@ -2469,13 +2506,19 @@
                         if (name instanceof Symbol) {
                             env.env[name.name] = arg;
                             break;
-                        } else if (name.car !== nil && arg.car !== nil) {
-                            env.env[name.car.name] = arg.car;
+                        } else if (name.car !== nil) {
+                            if (arg === nil) {
+                                env.env[name.car.name] = nil;
+                            } else {
+                                env.env[name.car.name] = arg.car;
+                            }
                         }
                         if (name.cdr === nil) {
                             break;
                         }
-                        arg = arg.cdr;
+                        if (arg !== nil) {
+                            arg = arg.cdr;
+                        }
                         name = name.cdr;
                     }
                     if (dynamic_scope) {
@@ -2485,12 +2528,12 @@
                     if (macro.cdr instanceof Pair) {
                         // this eval will return lips code
                         var rest = __doc__ ? macro.cdr.cdr : macro.cdr;
+                        if (macro_expand) {
+                            return rest.car;
+                        }
                         var pair = rest.reduce(function(result, node) {
                             return evaluate(node, { env, dynamic_scope, error });
                         });
-                        if (macro_expand) {
-                            return pair;
-                        }
                         // second evalute of code that is returned from macro
                         // need different env because we need to call it in scope
                         // were it was called
@@ -2671,18 +2714,21 @@
             the list with its value. Best used with macros but it can be used outside`),
         // ------------------------------------------------------------------
         clone: doc(function(list) {
+            typecheck('clone', list, 'pair');
             return list.clone();
         }, `(clone list)
 
             Function return clone of the list.`),
         // ------------------------------------------------------------------
         append: doc(function(list, item) {
+            typecheck('append', list, 'pair');
             return this.get('append!')(list.clone(), item);
         }, `(append list item)
 
             Function will create new list with value appended to the end. It return
             New list.`),
         'append!': doc(function(list, item) {
+            typecheck('append!', list, 'pair');
             if (isNull(item) || isEmptyList(item)) {
                 return list;
             }
@@ -2693,6 +2739,7 @@
              original list.`),
         // ------------------------------------------------------------------
         reverse: doc(function(arg) {
+            typecheck('reverse', arg, ['array', 'pair']);
             if (arg instanceof Pair) {
                 var arr = this.get('list->array')(arg).reverse();
                 return this.get('array->list')(arr);
@@ -2707,6 +2754,8 @@
             or array it will throw exception.`),
         // ------------------------------------------------------------------
         nth: doc(function(index, obj) {
+            typecheck('nth', index, 'number');
+            typecheck('nth', obj, ['array', 'pair']);
             if (obj instanceof Pair) {
                 var node = obj;
                 var count = 0;
@@ -2728,31 +2777,39 @@
             Function return nth element of the list or array. If used with different
             value it will throw exception`),
         // ------------------------------------------------------------------
-        list: doc(function() {
-            return Pair.fromArray([].slice.call(arguments));
+        list: doc(function(...args) {
+            return Pair.fromArray(args);
         }, `(list . args)
 
             Function create new list out of its arguments.`),
         // ------------------------------------------------------------------
         substring: doc(function(string, start, end) {
+            typecheck('substring', string, 'string');
+            typecheck('substring', start, 'number');
+            typecheck('substring', end, ['number', 'undefined']);
             return string.substring(start.valueOf(), end && end.valueOf());
         }, `(substring string start end)
 
             Function return part of the string starting at start ending with end.`),
         // ------------------------------------------------------------------
-        concat: doc(function() {
-            return [].join.call(arguments, '');
+        concat: doc(function(...args) {
+            args.forEach((arg, i) => typecheck('concat', arg, 'string', i + 1));
+            return args.join('');
         }, `(concat . strings)
 
             Function create new string by joining its arguments`),
         // ------------------------------------------------------------------
         join: doc(function(separator, list) {
+            typecheck('join', separator, 'string');
+            typecheck('join', list, 'pair');
             return this.get('list->array')(list).join(separator);
         }, `(join separator list)
 
             Function return string by joining elements of the list`),
         // ------------------------------------------------------------------
         split: doc(function(separator, string) {
+            typecheck('split', separator, ['regex', 'string']);
+            typecheck('split', string, 'string');
             return this.get('array->list')(string.split(separator));
         }, `(split separator string)
 
@@ -2760,12 +2817,18 @@
             be a string or regular expression.`),
         // ------------------------------------------------------------------
         replace: doc(function(pattern, replacement, string) {
+            typecheck('replace', pattern, ['regex', 'string']);
+            typecheck('replace', replacement, ['string', 'function']);
+            typecheck('replace', string, 'string');
             return string.replace(pattern, replacement);
         }, `(replace pattern replacement string)
 
-            Function change patter to replacement inside string.`),
+            Function change pattern to replacement inside string. Pattern can be string
+            or regex and replacement can be function or string.`),
         // ------------------------------------------------------------------
         match: doc(function(pattern, string) {
+            typecheck('match', pattern, ['regex', 'string']);
+            typecheck('match', string, 'string');
             var m = string.match(pattern);
             return m ? this.get('array->list')(m) : nil;
         }, `(match pattern string)
@@ -2773,6 +2836,8 @@
             function return match object from JavaScript as list.`),
         // ------------------------------------------------------------------
         search: doc(function(pattern, string) {
+            typecheck('search', pattern, ['regex', 'string']);
+            typecheck('search', string, 'string');
             return string.search(pattern);
         }, `(search pattern string)
 
@@ -2870,11 +2935,24 @@
         }, `(instanceof type obj)
 
             Function check of object is instance of object.`),
+        // ------------------------------------------------------------------
         'function?': doc(function(obj) {
             return typeof obj === 'function';
         }, `(function? expression)
 
             Function check if value is a function.`),
+        // ------------------------------------------------------------------
+        'real?': doc(function(value) {
+            if (type(value) !== 'number') {
+                return false;
+            }
+            if (value instanceof LNumber) {
+                return value.isFloat();
+            }
+            return LNumber.isFloat(value);
+        }, `(real? number)
+
+            Function check if value is real number.`),
         // ------------------------------------------------------------------
         'number?': doc(
             LNumber.isNumber,
@@ -2973,18 +3051,21 @@
             Display error message.`),
         // ------------------------------------------------------------------
         flatten: doc(function(list) {
+            typecheck('flatten', list, 'pair');
             return list.flatten();
         }, `(flatten list)
 
             Return shallow list from tree structure (pairs).`),
         // ------------------------------------------------------------------
         'array->list': doc(function(array) {
+            typecheck('array->list', array, 'array');
             return Pair.fromArray(array);
         }, `(array->list array)
 
             Function convert JavaScript array to LIPS list.`),
         // ------------------------------------------------------------------
         'list->array': doc(function(list) {
+            typecheck('list->array', list, ['pair', 'nil']);
             if (list instanceof Pair && list.isEmptyList()) {
                 return [];
             }
@@ -3006,11 +3087,11 @@
 
             Function convert LIPS list into JavaScript array.`),
         // ------------------------------------------------------------------
-        apply: doc(function(fn, args) {
+        apply: doc(function(fn, list) {
             typecheck('call', fn, 'function', 1);
-            typecheck('call', args, 'pair', 2);
-            return fn(...this.get('list->array')(args));
-        }, `(apply fn args)
+            typecheck('call', list, 'pair', 2);
+            return fn(...this.get('list->array')(list));
+        }, `(apply fn list)
 
             Function that call function with list of arguments.`),
         // ------------------------------------------------------------------
@@ -3058,31 +3139,37 @@
         }), `(try expr (catch (e) code)`),
         // ------------------------------------------------------------------
         find: doc(function find(arg, list) {
+            typecheck('find', arg, ['regex', 'function']);
+            typecheck('find', list, 'pair');
             if (isNull(list)) {
                 return nil;
             }
             var fn = matcher('find', arg);
             return unpromise(fn(list.car), function(value) {
-                if (value) {
+                if (value && value !== nil) {
                     return list.car;
                 }
                 return find(arg, list.cdr);
             });
-        }, `(Find fn list)
+        }, `(find fn list)
+            (find regex list)
 
-            Higher order Function find first value for which function
-            return true.`),
+            Higher order Function find first value for which function return true.
+            If called with regex it will create matcher function.`),
         // ------------------------------------------------------------------
-        'for-each': doc(function(fn, ...args) {
+        'for-each': doc(function(fn, ...lists) {
             typecheck('for-each', fn, 'function');
+            lists.forEach((arg, i) => {
+                typecheck('for-each', arg, ['pair', 'nil'], i + 1);
+            });
             // we need to use call(this because babel transpile this code into:
-            // var ret = map.apply(void 0, [fn].concat(args));
+            // var ret = map.apply(void 0, [fn].concat(lists));
             // it don't work with weakBind
-            var ret = this.get('map')(fn, ...args);
+            var ret = this.get('map')(fn, ...lists);
             if (isPromise(ret)) {
                 return ret.then(() => {});
             }
-        }, `(for-each fn . args)
+        }, `(for-each fn . lists)
 
             Higher order function that call function \`fn\` by for each
             value of the argument. If you provide more then one list as argument
@@ -3091,6 +3178,9 @@
         // ------------------------------------------------------------------
         map: doc(function map(fn, ...lists) {
             typecheck('map', fn, 'function');
+            lists.forEach((arg, i) => {
+                typecheck('map', arg, ['pair', 'nil'], i + 1);
+            });
             if (lists.some((x) => isEmptyList(x))) {
                 return nil;
             }
@@ -3109,6 +3199,8 @@
             returned by the call to map.`),
         // ------------------------------------------------------------------
         some: doc(function some(fn, list) {
+            typecheck('some', fn, 'function');
+            typecheck('some', list, ['pair', 'nil']);
             if (isNull(list)) {
                 return false;
             } else {
@@ -3123,6 +3215,10 @@
             return true. If it don't find the value it will return false`),
         // ------------------------------------------------------------------
         fold: doc(fold('fold', function(fold, fn, init, ...lists) {
+            typecheck('fold', fn, 'function');
+            lists.forEach((arg, i) => {
+                typecheck('fold', arg, ['pair', 'nil'], i + 1);
+            });
             if (lists.some(isEmptyList)) {
                 return init;
             }
@@ -3137,7 +3233,32 @@
              e.g. it call (fn a1 b1 (fn a2 b2 (fn a3 b3 '())))
              for: (fold fn '() alist blist`),
         // ------------------------------------------------------------------
+        pluck: doc(function(...keys) {
+            return function(obj) {
+                keys = keys.map(x => x instanceof Symbol ? x.name : x);
+                if (keys.length === 0) {
+                    return nil;
+                } else if (keys.length === 1) {
+                    const [key] = keys;
+                    return obj[key];
+                }
+                var result = {};
+                keys.forEach((key) => {
+                    result[key] = obj[key];
+                });
+                return result;
+            };
+        }, `(pluck . string)
+
+            If called with single string it will return function that will return
+            key from object. If called with more then one argument function will
+            return new object by taking all properties from given object.`),
+        // ------------------------------------------------------------------
         reduce: doc(fold('reduce', function(reduce, fn, init, ...lists) {
+            typecheck('reduce', fn, 'function');
+            lists.forEach((arg, i) => {
+                typecheck('reduce', arg, ['pair', 'nil'], i + 1);
+            });
             if (lists.some(isEmptyList)) {
                 return init;
             }
@@ -3154,12 +3275,14 @@
              for (reduce fn init alist blist`),
         // ------------------------------------------------------------------
         filter: doc(function(arg, list) {
+            typecheck('filter', arg, ['regex', 'function']);
+            typecheck('filter', list, ['pair', 'nil']);
             var array = this.get('list->array')(list);
             var result = [];
             var fn = matcher('filter', arg);
             return (function loop(i) {
                 function next(value) {
-                    if (value) {
+                    if (value && value !== nil) {
                         result.push(item);
                     }
                     return loop(++i);
@@ -3171,12 +3294,14 @@
                 return unpromise(fn(item, i), next);
             })(0);
         }, `(filter fn list)
+            (filter regex list)
 
             Higher order function that call \`fn\` for each element of the list
             and return list for only those elements for which funtion return
-            true value.`),
+            true value. If called with regex it will create matcher function.`),
         // ------------------------------------------------------------------
         range: doc(function(n) {
+            typecheck('range', n, 'number');
             if (n instanceof LNumber) {
                 n = n.valueOf();
             }
@@ -3477,10 +3602,10 @@
     // ----------------------------------------------------------------------
     ['floor', 'round', 'ceil'].forEach(fn => {
         global_env.set(fn, doc(function(value) {
+            typecheck(fn, value, 'number');
             if (value instanceof LNumber) {
                 return value[fn]();
             }
-            throw new Error(`${typeof value} ${value.toString()} is not a number`);
         }, `(${fn} number)
 
             Function calculate ${fn} of a number.`));
@@ -3552,12 +3677,20 @@
         if (position !== null) {
             postfix += ` argument ${position}`;
         }
+        if (expected instanceof Array) {
+            const last = expected[expected.length - 1];
+            expected = expected.slice(0, -1).join(', ') + ' or ' + last;
+        }
         return `Expecting ${expected} got ${got}${postfix}`;
     }
     // ----------------------------------------------------------------------
     function typecheck(fn, arg, expected, position = null) {
         const arg_type = type(arg);
-        if (arg_type !== expected) {
+        var match = false;
+        if (expected instanceof Array && expected.includes(arg_type)) {
+            match = true;
+        }
+        if (!match && arg_type !== expected) {
             throw new Error(typeErrorMessage(fn, arg_type, expected, position));
         }
     }
@@ -3795,7 +3928,9 @@
                     dynamic_scope,
                     error: (e, code) => {
                         if (code) {
-                            e.code = code.toString();
+                            // LIPS stack trace
+                            e.code = e.code || [];
+                            e.code.push(code.toString());
                         }
                         throw e;
                     }