+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.12.0
+### Features
+* new macro `begin*` that run their arguments in parallel if they async
+* add typecheck to rest of the functions
+* new `real?` and `pluck` functions
+* stack trace for exceptions
+### Bug fixes
+* fix `find` and `filter` to check if function from argument return nil
+* fix accessing FileReader object (try to copy read only prototype)
+* fix parser error when parsing symbol macros [#10](https://github.com/jcubic/lips/issues/10)
+* fix undefined values in macro with missing arguments [#11](https://github.com/jcubic/lips/issues/11)
+* fix macroexpand [#9](https://github.com/jcubic/lips/issues/9)
+* fix line breaking in S-Expression Formatter
+
 ## 0.11.2
 ### Breaking
 * remove `nop` (you can use `(begin undefined)`)